feat: string extensions
changes:
- added `compress` method
- added `decompress` method
- added unit tests
- some refactoring

closes #36
